requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
accelerate=0.21.0=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
aliyun-python-sdk-core=2.13.36=pypi_0
aliyun-python-sdk-kms=2.16.1=pypi_0
ansi2html=1.8.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.7.1=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
bert-score=0.3.13=pypi_0
bitsandbytes=0.41.1=pypi_0
black=23.7.0=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
brotlipy=0.7.0=py38h27cfd23_1003
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.1=pypi_0
cchardet=2.1.7=pypi_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py38h5eee18b_3
chardet=5.2.0=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.1.6=pypi_0
cmake=3.27.0=pypi_0
colorama=0.4.6=pypi_0
colorlog=6.7.0=pypi_0
comm=0.1.4=pypi_0
configargparse=1.7=pypi_0
contourpy=1.1.0=pypi_0
cosine-annealing-warmup=2.0=pypi_0
crcmod=1.7=pypi_0
cryptography=41.0.2=py38h22a60cf_0
cudatoolkit=11.7.0=hd8887f6_10
cycler=0.11.0=pypi_0
dash=2.13.0=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
datasets=2.14.3=pypi_0
debugpy=1.6.7.post1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
descartes=1.1.0=pypi_0
dill=0.3.7=pypi_0
docker-pycreds=0.4.0=pypi_0
evaluate=0.4.0=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.18.0=pypi_0
filelock=3.12.2=pypi_0
fire=0.5.0=pypi_0
flake8=6.1.0=pypi_0
flask=2.2.5=pypi_0
fonttools=4.42.0=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
gcc-6=6.1.0=2
gitdb=4.0.10=pypi_0
gitpython=3.1.32=pypi_0
gmp=6.1.0=1
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.56.2=pypi_0
huggingface-hub=0.16.4=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.4=py38h06a4308_0
imageio=2.31.1=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.0=pypi_0
iniconfig=2.0.0=pypi_0
inquirerpy=0.3.4=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.25.1=pypi_0
ipython=8.12.2=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.0=pypi_0
isl=0.17.1=0
isoduration=20.11.0=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.0=pypi_0
jinja2=3.1.2=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.3.1=pypi_0
json5=0.9.14=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.19.0=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.3.1=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.3.1=pypi_0
jupyter-events=0.7.0=pypi_0
jupyter-lsp=2.2.0=pypi_0
jupyter-server=2.7.2=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=4.0.5=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.24.0=pypi_0
jupyterlab-widgets=3.0.8=pypi_0
kiwisolver=1.4.4=pypi_0
language-evaluation=0.1.0=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lightning-utilities=0.9.0=pypi_0
line-profiler=4.0.3=pypi_0
lit=16.0.6=pypi_0
llvmlite=0.31.0=pypi_0
lyft-dataset-sdk=0.0.8=pypi_0
markdown=3.4.4=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mccabe=0.7.0=pypi_0
mdurl=0.1.2=pypi_0
mistune=2.0.5=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.6=py38h417a72b_1
mkl_random=1.2.2=py38h417a72b_1
mmcv=1.4.0=pypi_0
mmdet=2.14.0=pypi_0
mmengine=0.8.4=pypi_0
mmsegmentation=0.14.1=pypi_0
model-index=0.1.11=pypi_0
more-itertools=10.1.0=pypi_0
mpc=1.0.1=0
mpfr=3.1.2=0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclient=0.8.0=pypi_0
nbconvert=7.4.0=pypi_0
nbformat=5.5.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.7=pypi_0
networkx=2.2=pypi_0
nltk=3.8.1=pypi_0
notebook=7.0.2=pypi_0
notebook-shim=0.2.3=pypi_0
numba=0.48.0=pypi_0
numpy=1.24.4=pypi_0
nuscenes-devkit=1.1.10=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
open3d=0.16.0=pypi_0
opencv-python=4.8.0.74=pypi_0
opendatalab=0.0.10=pypi_0
openmim=0.3.9=pypi_0
openssl=3.0.10=h7f8727e_2
openxlab=0.0.22=pypi_0
ordered-set=4.1.0=pypi_0
oss2=2.17.0=pypi_0
overrides=7.4.0=pypi_0
packaging=23.1=pypi_0
pandas=1.4.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.2=pypi_0
pathtools=0.1.2=pypi_0
peft=0.4.0=pypi_0
pexpect=4.8.0=pypi_0
pfzy=0.3.4=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.0.0=pypi_0
pip=23.2.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.10.0=pypi_0
plotly=5.16.1=pypi_0
pluggy=1.3.0=pypi_0
plyfile=1.0.1=pypi_0
pooch=1.4.0=pyhd3eb1b0_0
prettytable=3.8.0=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.23.4=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=12.0.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.7=pypi_0
pycodestyle=2.11.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.18.0=pypi_0
pydeprecate=0.3.2=pypi_0
pyflakes=3.1.0=pypi_0
pygments=2.16.1=pypi_0
pyopenssl=23.2.0=py38h06a4308_0
pyparsing=3.0.9=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=py38h06a4308_0
pytest=7.4.0=pypi_0
python=3.8.17=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
pytorch-lightning=1.7.0=pypi_0
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.1=pypi_0
qtconsole=5.4.3=pypi_0
qtpy=2.4.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
regex=2023.6.3=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
retrying=1.3.4=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.4.2=pypi_0
rouge-score=0.1.2=pypi_0
rpds-py=0.10.0=pypi_0
rsa=4.9=pypi_0
safetensors=0.3.1=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.7.3=pypi_0
send2trash=1.8.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=1.29.2=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=60.2.0=pypi_0
shapely=1.8.5=pypi_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h5eee18b_0
src=0.0.0=dev_0
stack-data=0.6.2=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.3=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=pypi_0
terminaltables=3.1.10=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.7.10=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=pypi_0
torch=2.0.1=pypi_0
torchaudio=2.0.2=pypi_0
torchmetrics=0.11.1=pypi_0
torchvision=0.15.2=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.31.0=pypi_0
trimesh=2.35.39=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023.3=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.0.4=pypi_0
wandb=0.15.8=pypi_0
wcwidth=0.2.6=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.2=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py38h06a4308_0
widgetsnbextension=4.0.8=pypi_0
xxhash=3.3.0=pypi_0
xz=5.4.2=h5eee18b_0
yapf=0.40.1=pypi_0
yarl=1.9.2=pypi_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
